bug: 키보드 올라오고 내려가는 것에 따른 view 변화 추가 (#77Share)
